requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
pydantic==2.10.3
python-socketio